+++ Changes.md
@@ -1,6 +1,15 @@
 Change log
 ===========
 
+# v 0.30.3
+## Changed
+* static下のファイル構成を/static/files.jsonファイルとして出力し環境変数化しない変更
+* 静的サイト化テスト
+
+# v 0.30.2
+## Fixed
+* staticを読むAPIを除去
+
 # v 0.30.1
 ## Fixed
 * アバターがemoを反映する修正
@@ -9,3 +18,4 @@ Change log
 
 ## TODO
 * 心の声の表示法
+* reactivityの実装（contextに含まれるので不要かも）